+++ lib/screens/sign_in.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:grantconsent/utilities/constants.dart';
-import 'package:grantconsent/utilities/styles.dart';
-import 'package:grantconsent/utilities/custom_widgets.dart';
 import 'package:grantconsent/screens/sign_up.dart';
+import 'package:grantconsent/utilities/constants.dart';
+import 'package:grantconsent/utilities/custom_widgets.dart';
+import 'package:grantconsent/utilities/styles.dart';
+
 class SignIn extends StatefulWidget {
   @override
   _SignInState createState() => _SignInState();
@@ -10,6 +11,7 @@ class SignIn extends StatefulWidget {
 
 class _SignInState extends State<SignIn> {
   bool checkBoxValue = false;
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -26,11 +28,9 @@ class _SignInState extends State<SignIn> {
                   Column(
                     children: <Widget>[
                       Padding(
-                        padding: const EdgeInsets.only(
-                            left: 35.0, right: 35.0, top: 35.0, bottom: 5),
-                        child:GrantConsentLogo(LogoType.largeWithText)
-                      ),
-
+                          padding: const EdgeInsets.only(
+                              left: 35.0, right: 35.0, top: 35.0, bottom: 5),
+                          child: GrantConsentLogo(LogoType.mediumWithText)),
                     ],
                   ),
                 ],
@@ -68,7 +68,6 @@ class _SignInState extends State<SignIn> {
                         children: <Widget>[
                           Row(children: <Widget>[
                             Checkbox(
-
                                 value: checkBoxValue,
                                 activeColor: kButtonTextColor2,
                                 onChanged: (bool newValue) {
@@ -88,38 +87,32 @@ class _SignInState extends State<SignIn> {
                         ],
                       ),
                     ),
-
                     Padding(
                       padding: const EdgeInsets.only(top: 50.0),
                       child: Column(
                         children: <Widget>[
                           Padding(
                             padding: const EdgeInsets.only(bottom: 15.0),
-                            child:  UserActionButton(
-                              onTap: () {
-                                Navigator.push(context,
-                                    MaterialPageRoute(builder: (context) => SignIn()));
-
-                              },
-                              label: 'Sign In',
-                              filled: true
-
-                            ),
+                            child: UserActionButton(
+                                onTap: () {
+                                  Navigator.push(
+                                      context,
+                                      MaterialPageRoute(
+                                          builder: (context) => SignIn()));
+                                },
+                                label: 'Sign In',
+                                filled: true),
                           ),
-                          Text(
-                            "OR",
-                            style:kOrTextStyle
-                          ),
+                          Text("OR", style: kOrTextStyle),
                           Padding(
                             padding: const EdgeInsets.only(top: 15.0),
-                            child:  UserGoogleButton(
+                            child: UserGoogleButton(
                               onTap: () {
-                                Navigator.push(context,
-                                    MaterialPageRoute(builder: (context) => SignIn()));
-
+                                Navigator.push(
+                                    context,
+                                    MaterialPageRoute(
+                                        builder: (context) => SignIn()));
                               },
-
-
                             ),
                           ),
                         ],
@@ -138,16 +131,16 @@ class _SignInState extends State<SignIn> {
               children: <Widget>[
                 Text(
                   'Don\'t have an account? ',
-                  style: TextStyle(color: kButtonTextColor2, fontSize: 15),
+                  style: TextStyle(color: kButtonTextColor2, fontSize: 13),
                 ),
                 GestureDetector(
-                  onTap: (){
+                  onTap: () {
                     Navigator.push(context,
                         MaterialPageRoute(builder: (context) => SignUp()));
                   },
                   child: Text(
                     'Sign Up',
-                    style: TextStyle(color: Colors.white, fontSize: 15),
+                    style: TextStyle(color: Colors.white, fontSize: 13),
                   ),
                 )
               ],
@@ -158,4 +151,3 @@ class _SignInState extends State<SignIn> {
     );
   }
 }
-
